Convert datetime args to localtime
In order to compare whether the available event falls within the
selected start and end dates, the timestamp of the event needs to be
compared to the selected start and end dates.

This commit adds localization to the datetimes entered as CLI arguments.
It also converts string input from the CLI arguments to datetime
objects.

As part of this process, the processing of the arguments was improved
to show the default settings in ARG_DEFAULTS in the parser's output

Signed off by: Rod Manning <rod.t.manning@gmail.com>