Add create_blocks() function
This commit added detail to the create_blocks function which is used to
populate the list of blocks that we're compiling free/busy time for.

This function is a fairly simple iterator that keeps adding blocks until
> the nominated end datetime for the calendar that's being generated.

The function also does a bit of trimming to set the last block in the day
to the last available hour in the day.

Couple of other minor changes:

- Put the program logic above the variables in the main docstring (seems
  to make more sense to know how it works before working out what variables
  are required).

- Changed the "day_start" and "day_end" args to ints (e.g. "06" = 6 AM)

- Changed the Event and Block object __init__ functions to use **kwargs
  instead of named args (bit more flexible).

- Renamed generic lists used to hold things as "stack" (consistency :) )

- Added code to convert args that should be int to actually be int.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>