Add code to assign busy time to each block
This commit adds simple code to assign free/busy time to a block for
each event.

The logic is as follows:

- For each event, iterate through *all* the blocks and check whether the
  event and block overlap.

  - If they overlap, calculate the size of the overlap and update the
    blocks free busy time.

  - If they don't overlap then don't do anything.

The check_overlap() and calculate_overlap() functions were added to perform
these computations, whilst the assign_hours_to_blocks() function is now
a simple flow control function.

As assign() method was added to the Block object to assist with this
process.

This commit also includes code to fix errors in the creation of the
Block objects that was causing it to skip days. The error was in the logic
which was replacing the hour and then incrementing 24 hours, which wasn't
working if the block ended at midnight.

The new code calculates the day_start independently, and just keeps
incrementing this 24 hours whenever a new day is detected.

Finally, added validation to the --start and --end arguments to check
reasonableness, and added a 'length' property to the Event object.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>