Add additional detail to the main docstring
The detail in the main docstring explains what the program does, and that
it effectively just parses data to plug in to some other program/site.

Provided detail in the main docstring about the properties of the Block
objects that the program returns.

Also replaced the test CalDAV file code with real code to retrieve the
actual calendar from the server.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>