Add test suite (initial commit)
Added a test.py file to contain the test suite. Currently, this contains
a unittest for the creation of a Block. Further work required to expand
to cover all areas of the code.

This required renaming the main file to replace the '-' with an '_' (as
'-' isn't callable from Python (whoops!).

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>