Added the Realm as a default option
Basically, forgot what realm to use and spent 45 minutes with telnet
trying to find it out :/

Simpler just to leave it as a default under VC, as minimal security risk
given that it's sent as plain text in the HTTP header response.

Also re-named the main() loop to get_availability() so that it can be
imported from within other scripts more easily.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>