Add json output to std out
The script needs a way to get data out :)

It could be called from python (in which case we can return an object--
that will be the next edit :) ).

However, it also makes sense to output the data as json to be deserialized
back into a site or page.

To enable this, added an argparse option to determine how the output is
returned and added some basic logic for this in the main loop.

Also added a 'as_json()' method to Block objects, and a _json_default()
method to serialize datetime and timedelta objects as strings.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>