Add Error Checking to Ensure Invalid Hours Handled
If the amount of work assigned in a block exceeded the limits of the
block then the program failed but no error was raised.

Expected output is for the program to recognise this situation and
adjust the free/busy time to show 0 free and 100% busy.

Added error checking code to the assign() method for Block objects to
detect these cases and handle them as described above. (Needs to go into
the test cases!).

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>