Update Caldav for new KolabNow Calendar Format
With updates to KolabNow/Roundcube calendar, the format of returned date
and datetime strings for the DTEND/DTSTART was different. Updated code
to catch new format of these input strings.

Also, improved handling of 'all day' events by now casting them to a
fully localized datetime object, and then normalizing them to
UTC. Creates a consistent data format for later processing, and appears
to avoid having to have hacks like -1 day in order for the output to be
calculated correctly.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>